Add button color variable, site wide consistency
It is possible to set a custom color for buttons, dropdowns and input
fields. In addition, more of the site has been themed properly, this
includes the settings page for repos. Code that has been added/deleted
in commits shows up properly with its color but is still a work in
progress.